Lab 3
This lab report for Practical 3 of the "Computer Intrusion Detection" course focuses on gathering information about a target webserver using web-based footprinting techniques. The target selected was Watsons Malaysia (https://www.watsons.com.my/). The lab explains the purpose of a nameserver, which maps domain names to IP addresses and handles DNS records. It also discusses IP delegation, a method of assigning blocks of IP addresses to other organizations. The report identifies a subdomain from the webserver and explores the role of F5 BIG-IP, a tool used for load balancing, traffic management, and web application security. This lab practical introduces core web footprinting techniques that are essential for penetration testing and reconnaissance.

Objectives from the Lab:
1. Identify and analyze the target webserver’s basic information.
2. Understand the function and importance of nameservers in DNS resolution.
3. Explain the concept and purpose of IP delegation.
4. Discover subdomains associated with the target server.
5. Understand the use of F5 BIG-IP in network traffic management and security.